requirements.txt
google-genai==1.47.0
google-adk==1.17.0
crewai==0.134.0
pydantic==2.12.3
pydantic-settings==2.10.1
fastapi==0.115.13
uvicorn>=0.34.2
mcp==1.20.0
pymongo==4.13.2
streamlit==1.51.0
python-dotenv==1.1.1
loguru==0.7.3
openai==1.90.0
litellm==1.79.1